requirements.txt
tensorflow==2.13.0
boto3==1.28.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
PyYAML==6.0.1
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1

# SageMaker
sagemaker==2.175.0

# For data processing
pyarrow==12.0.1

# For testing
pytest==7.4.0
pytest-cov==4.1.0
